# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "PERX Queue API",
        "title": "PERX Queue API",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
        "/task/add": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "Add task",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TaskDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "success response",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "failure response",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "422": {
                        "description": "failure response",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/task/list": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "Get task list",
                "responses": {
                    "200": {
                        "description": "success response",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.TaskModel"
                            }
                        }
                    },
                    "400": {
                        "description": "failure response",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.TaskDTO": {
            "type": "object",
            "required": [
                "I",
                "TTL",
                "d",
                "n",
                "n1"
            ],
            "properties": {
                "I": {
                    "type": "number",
                    "example": 5.5
                },
                "TTL": {
                    "type": "number",
                    "example": 120
                },
                "d": {
                    "type": "number",
                    "example": 11.5
                },
                "n": {
                    "type": "integer",
                    "example": 5
                },
                "n1": {
                    "type": "number",
                    "example": 21.33
                }
            }
        },
        "models.TaskModel": {
            "type": "object",
            "properties": {
                "I": {
                    "type": "number",
                    "example": 5.5
                },
                "TTL": {
                    "type": "number",
                    "example": 120
                },
                "current_iteration": {
                    "type": "integer",
                    "example": 3
                },
                "current_value": {
                    "type": "number",
                    "example": 21.33
                },
                "d": {
                    "type": "number",
                    "example": 11.5
                },
                "n": {
                    "type": "integer",
                    "example": 5
                },
                "n1": {
                    "type": "number",
                    "example": 21.33
                },
                "queue_index": {
                    "type": "integer",
                    "example": 0
                },
                "task_created_at": {
                    "type": "string",
                    "example": "2022-01-01T00:00:00+00:00"
                },
                "task_finished_at": {
                    "type": "string",
                    "example": "2022-01-01T00:00:00+00:00"
                },
                "task_id": {
                    "type": "integer",
                    "example": 1
                },
                "task_started_at": {
                    "type": "string",
                    "example": "2022-01-01T00:00:00+00:00"
                },
                "task_status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.TaskStatus"
                        }
                    ],
                    "example": "in_queue"
                }
            }
        },
        "models.TaskStatus": {
            "type": "string",
            "enum": [
                "in queue",
                "in progress",
                "completed"
            ],
            "x-enum-varnames": [
                "statusInQueue",
                "statusInProgress",
                "statusCompleted"
            ]
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "response detail"
                }
            }
        }
    }
}